updated: restructuring geocogs
